move remove exposition of DatabaseReference in UserRepository
