+++ app/src/main/java/ch/epfl/sdp/blindly/database/UserRepository.kt
@@ -11,27 +11,13 @@ import ch.epfl.sdp.blindly.user.User.Companion.toUser
 import ch.epfl.sdp.blindly.user.storage.UserCache
 import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.Query
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 import kotlin.reflect.KSuspendFunction1
 
-/**
- * The main access to Firebase firestore
- *
- * @property db the instance of FirebaseFirestore
- * @property userCache the local cache
- */
-class UserRepository @Inject constructor(
-    private val db: FirebaseFirestore,
-    private val userCache: UserCache
-) {
-
-    companion object {
-        private const val TAG = "UserRepository"
-        private const val USER_COLLECTION: String = "usersMeta"
-    }
-
+interface UserRepository {
     /**
      * Given a uid, if the user is cached locally return this user, otherwise
      * look for the user in firestore and update the cache
@@ -39,14 +25,7 @@ class UserRepository @Inject constructor(
      * @param uid the uid of the user to retrieve
      * @return the user with the corresponding uid or null if they doesn't exist
      */
-    suspend fun getUser(uid: String): User? {
-        val cached: User? = userCache.get(uid)
-        if (cached != null) {
-            Log.d(TAG, "Found user with uid: $uid in cache")
-            return cached
-        }
-        return refreshUser(uid)
-    }
+    suspend fun getUser(uid: String): User?
 
     /**
      * Look for the user with the corresponding uid in firestore and store it in the local cache
@@ -54,31 +33,7 @@ class UserRepository @Inject constructor(
      * @param uid the uid of the user to retrieve in firestore
      * @return the user with the corresponding uid or null if he/she/it doesn't exist
      */
-    suspend fun refreshUser(uid: String): User? {
-        return try {
-            val freshUser = db.collection(USER_COLLECTION)
-                .document(uid).get().await().toUser()
-            if (freshUser != null) {
-                Log.d(TAG, "Put User \"$uid\" in local cache")
-                userCache.put(uid, freshUser)
-            }
-            Log.d(TAG, "Retrieve User \"$uid\" in firestore")
-            return freshUser
-        } catch (e: Exception) {
-            Log.e(TAG, "Error getting user details", e)
-            null
-        }
-    }
-
-    private suspend fun <T> updateLocalCache(uid: String, field: String, newValue: T) {
-        val user = userCache.get(uid)
-        if (user != null) {
-            Log.d(TAG, "Updated user in local cache")
-            userCache.put(uid, User.updateUser(user, field, newValue))
-        } else {
-            refreshUser(uid)
-        }
-    }
+    suspend fun refreshUser(uid: String): User?
 
     /**
      * Update a given field of the user's information (and call refreshUser to update or set the
@@ -89,64 +44,27 @@ class UserRepository @Inject constructor(
      * @param newValue the new value to set for the user
      */
 
-    suspend fun <T> updateProfile(uid: String, field: String, newValue: T) {
-        if (newValue !is String && newValue !is List<*> && newValue !is Int)
-            throw IllegalArgumentException("Expected String, List<String> or Int")
-
-        db.collection(USER_COLLECTION)
-            .document(uid)
-            .update(field, newValue)
-        Log.d(TAG, "Updated user")
-        //Put updated value into the local cache
-        updateLocalCache(uid, field, newValue)
-    }
+    suspend fun <T> updateProfile(uid: String, field: String, newValue: T)
 
     /**
-     * Get the collection reference of the database of users.
+     * Asynchronously get the users
      *
-     * @return the reference of the database
+     * @return The list of the matching users for a query
      */
-    fun getCollectionReference(): CollectionReference {
-        return db.collection(USER_COLLECTION)
-    }
+    suspend fun query(query: Query): List<User>
 
     @RequiresApi(Build.VERSION_CODES.N)
     suspend fun getMyMatches(
         viewLifecycleOwner: LifecycleOwner,
         userId: String,
         setupAdapter: KSuspendFunction1<ArrayList<MyMatch>, Unit>
-    ) {
-        var myMatchesUids: List<String>
-        var myMatches: ArrayList<MyMatch>?
-        val docRef = getCollectionReference().document(userId)
-        docRef.addSnapshotListener { snapshot, e ->
-            if (e != null) {
-                Log.w(TAG, "Listen failed.", e)
-                return@addSnapshotListener
-            }
+    )
 
-            if (snapshot != null && snapshot.exists()) {
-                Log.d(TAG, "Current data: ${snapshot.data}")
-                myMatchesUids = snapshot["matches"] as List<String>
-                viewLifecycleOwner.lifecycleScope.launch {
-                    myMatches = arrayListOf()
-                    for (userId in myMatchesUids) {
-                        myMatches!!.add(
-                            MyMatch(
-                                getUser(userId)?.username!!,
-                                userId,
-                                false
-                            )
-                        )
-                    }
-                    setupAdapter(myMatches!!)
-
-                }
-            } else {
-                Log.d(TAG, "Current data: null")
-            }
-        }
-
-    }
-
+    /**
+     * Query for the user repositiry
+     *
+     * @property passions the passions the user must have
+     * @property gender the gender the user must have
+     */
+    class Query(var passions: List<String>? = null, var gender: String? = null )
 }
